decoding binary data for a text/json responses

+++ eoxserver/processing/gdal/reftools.py
@@ -240,8 +240,8 @@ class CoordinateTransformation(object):
 
     def __init__(self, src_srs, dst_srs):
         self._handle = OCTNewCoordinateTransformation(
-            C.cast(long(src_srs.this), C.c_void_p),
-            C.cast(long(dst_srs.this), C.c_void_p)
+            C.cast(int(src_srs.this), C.c_void_p),
+            C.cast(int(dst_srs.this), C.c_void_p)
         )
 
     @property
@@ -255,7 +255,7 @@ class CoordinateTransformation(object):
 def _create_referenceable_grid_transformer(ds, method, order):
     # TODO: check method and order
     num_gcps = ds.GetGCPCount()
-    gcps = GDALGetGCPs(C.cast(long(ds.this), C.c_void_p))
+    gcps = GDALGetGCPs(C.cast(int(ds.this), C.c_void_p))
     handle = None
     if method == METHOD_GCP:
         handle = GDALCreateGCPTransformer(num_gcps, C.c_void_p(gcps), order, 0)
@@ -356,11 +356,11 @@ def _create_generic_transformer(src_ds, src_wkt, dst_ds, dst_wkt, method, order)
     # TODO: check method and order
 
     try:
-        src_ds = C.c_void_p(long(src_ds.this))
+        src_ds = C.c_void_p(int(src_ds.this))
     except AttributeError:
         pass
     try:
-        dst_ds = C.c_void_p(long(dst_ds.this))
+        dst_ds = C.c_void_p(int(dst_ds.this))
     except AttributeError:
         pass
 
@@ -625,7 +625,7 @@ def create_rectified_vrt(path_or_ds, vrt_path, srid=None,
         raise ValueError('size and resolution ar mutually exclusive')
 
     ds = _open_ds(path_or_ds)
-    ptr = C.c_void_p(long(ds.this))
+    ptr = C.c_void_p(int(ds.this))
 
     if srid:
         srs = osr.SpatialReference()
@@ -656,7 +656,7 @@ def create_rectified_vrt(path_or_ds, vrt_path, srid=None,
     # options.eResampleAlg = resample
     # options.pfnTransformer = GDALGenImgProjTransform
     # options.pTransformerArg = transformer
-    # options.hDstDS = C.c_void_p(long(ds.this))
+    # options.hDstDS = C.c_void_p(int(ds.this))
 
     # nb = options.nBandCount = ds.RasterCount
